Starting to understand why monads are functors
On my own; as a proof, I implement map in terms of flatMap. Doing so I
proof that having a flatMap function (as a Monad by definition has) it
also has a derived map operation that conforms to the map function
signature defined in the Functor trait. We can say as a conclusion that
any Monad is also a Functor.